project/requirements.txt
Pillow==3.1.1
matplotlib==1.5.1
numpy==1.10.4
scikit-learn==0.17.1
scikit-image==0.12.3
